requirements.txt
wandb~=0.19.8
numpy~=2.2.3
datasets~=3.3.2
scikit-learn~=1.6.1
transformers~=4.51.3
huggingface-hub~=0.30.2
pandas~=2.2.3
nltk~=3.9.1
tqdm~=4.67.1
torch~=2.6.0
langchain-text-splitters~=0.3.6
python-dotenv~=1.0.1
evaluate~=0.4.3